+++ .prettierrc.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+  singleQuote: true,
+  semi: false,
+  trailingComma: "all",
+  tabWidth: 2,
+  printWidth: 120,
+};