+++ .prettierrc.js
@@ -1,7 +1,0 @@
-module.exports = {
-  singleQuote: true,
-  semi: false,
-  trailingComma: 'all',
-  tabWidth: 2,
-  printWidth: 120,
-}