+++ jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  clearMocks: true,
-  // setupFilesAfterEnv: ['regenerator-runtime/runtime'],
-  testPathIgnorePatterns: ['/node_modules/'],
-}